Rename CommCareResourceManager to ResourceManager
